[BE/YN]공지사항 api에 jwt 적용

See merge request s07-bigdata-recom-sub2/S07P22B306!14

+++ backend/mgmg-project/src/main/java/com/ssafy/mgmgproject/api/controller/NoticeController.java
@@ -5,8 +5,10 @@ import com.ssafy.mgmgproject.api.response.NoticeListMapping;
 import com.ssafy.mgmgproject.api.response.NoticeListResponse;
 import com.ssafy.mgmgproject.api.response.NoticeResponse;
 import com.ssafy.mgmgproject.api.service.NoticeService;
+import com.ssafy.mgmgproject.api.service.UserService;
 import com.ssafy.mgmgproject.common.model.response.BaseResponseBody;
 import com.ssafy.mgmgproject.db.entity.Notice;
+import com.ssafy.mgmgproject.db.entity.User;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiResponse;
 import io.swagger.annotations.ApiResponses;
@@ -16,6 +18,7 @@ import org.springframework.data.domain.Sort;
 import org.springframework.data.web.PageableDefault;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
 import springfox.documentation.annotations.ApiIgnore;
 
@@ -30,6 +33,9 @@ public class NoticeController {
     @Autowired
     private NoticeService noticeService;
 
+    @Autowired
+    private UserService userService;
+
     @PostMapping()
     @ApiOperation(value = "공지사항 작성", notes = "공지사항 제목과 내용을 작성한다.")
     @ApiResponses({
@@ -38,8 +44,15 @@ public class NoticeController {
             @ApiResponse(code = 500, message = "서버 오류", response = BaseResponseBody.class)
     })
     public ResponseEntity<? extends BaseResponseBody> writeNotice(@ApiIgnore Authentication authentication, @RequestBody NoticeRequest noticeRequest) throws Exception{
-        // JWT 관리자 체크
         try {
+            UserDetails userDetails = (UserDetails)authentication.getDetails();
+            String userId = userDetails.getUsername();
+            User user = userService.getByUserId(userId);
+
+            if(!user.isAdmin()){
+                return ResponseEntity.status(403).body(BaseResponseBody.of(403, "관리자만 접근 가능합니다."));
+            }
+
             Notice notice = noticeService.writeNotice(noticeRequest);
             if(notice!=null) {
                 return ResponseEntity.status(200).body(NoticeResponse.of(notice, 200, "공지사항이 등록되었습니다."));
@@ -59,11 +72,8 @@ public class NoticeController {
             @ApiResponse(code = 500, message = "서버 오류", response = BaseResponseBody.class)
     })
     public ResponseEntity<? extends BaseResponseBody> listNotice(
-            @ApiIgnore Authentication authentication,
             @RequestParam(required = false) String keyword,
             @PageableDefault(sort = "noticeDate",  direction = Sort.Direction.DESC) Pageable pageable){
-        // JWT
-
         Map<String,Object> result = noticeService.selectNoticeList(keyword,pageable);
         if(result==null){
             return ResponseEntity.status(401).body(BaseResponseBody.of(402, "공지사항 목록이 없습니다."));
@@ -79,9 +89,7 @@ public class NoticeController {
             @ApiResponse(code = 401, message = "공지사항 상세정보 조회 실패", response = BaseResponseBody.class),
             @ApiResponse(code = 402, message = "해당 공지 없음", response = BaseResponseBody.class),
             @ApiResponse(code = 500, message = "서버 오류", response = BaseResponseBody.class)})
-    public ResponseEntity<? extends BaseResponseBody> detailNotice(@ApiIgnore Authentication authentication, @PathVariable long noticeNo){
-        // JWT
-
+    public ResponseEntity<? extends BaseResponseBody> detailNotice(@PathVariable long noticeNo){
         Notice notice = noticeService.getByNoticeNo(noticeNo);
         if (notice == null) return ResponseEntity.status(402).body(BaseResponseBody.of(402, "해당하는 공지사항이 없습니다."));
         return ResponseEntity.status(200).body(NoticeResponse.of(notice, 200, "공지사항 상세조회를 성공하였습니다."));
@@ -97,7 +105,13 @@ public class NoticeController {
             @ApiResponse(code = 500, message = "서버 오류", response = BaseResponseBody.class)
     })
     public ResponseEntity<? extends BaseResponseBody> updateNotice(@ApiIgnore Authentication authentication, @PathVariable long noticeNo, @RequestBody NoticeRequest noticeRequest) throws Exception{
-        // JWT 관리자 체크
+        UserDetails userDetails = (UserDetails)authentication.getDetails();
+        String userId = userDetails.getUsername();
+        User user = userService.getByUserId(userId);
+
+        if(!user.isAdmin()){
+            return ResponseEntity.status(403).body(BaseResponseBody.of(403, "관리자만 접근 가능합니다."));
+        }
         try {
             Notice notice = noticeService.updateNotice(noticeNo,noticeRequest);
             if(notice!=null) {
@@ -119,7 +133,13 @@ public class NoticeController {
             @ApiResponse(code = 500, message = "서버 오류", response = BaseResponseBody.class)
     })
     public ResponseEntity<? extends BaseResponseBody> deleteNotice(@ApiIgnore Authentication authentication, @PathVariable long noticeNo) throws Exception{
-        // JWT 관리자 체크
+        UserDetails userDetails = (UserDetails)authentication.getDetails();
+        String userId = userDetails.getUsername();
+        User user = userService.getByUserId(userId);
+
+        if(!user.isAdmin()){
+            return ResponseEntity.status(403).body(BaseResponseBody.of(403, "관리자만 접근 가능합니다."));
+        }
         boolean result;
         try {
             result = noticeService.deleteNotice(noticeNo);
